requirements.txt
flask
flask-cors
pdfplumber
pygithub
gunicorn